Refactor code
Moved running function to its own file RunBash.Java
See #19 for more information
